feat(payments): payment detail (sheet + inline pane) with per-project breakdown, fundraiser link, dedication
